+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- *Breaking*: renamed `StandingOrderReoccurenceType` to `StandingOrderReoccurrenceType`
 - Add method to fetch account stats (`models.account.getStats`)
 
 ## [0.27.0] - 2020-04-20